pyproject.toml
[tool.poetry]
name = "c-cb"
version = "0.1.0"
description = ""
authors = ["yan-shen.lee.e <yan-shen.lee.e@thalesdigital.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
llama-index-core = "^0.12.6"
llama-index-readers-file = "^0.4.1"
llama-index-llms-ollama = "^0.5.0"
llama-index-embeddings-huggingface = "^0.4.0"
llama-index-vector-stores-chroma = "^0.4.1"
tree-sitter = "^0.23.2"
tree-sitter-kotlin = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
